+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Parse and highlight format strings
 - Parse and highlight interfaces
 - Add option to use specific `.clang-format` file for AngelScript
+- Focus Rider when hitting breakpoints